# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Agency API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface Trip {
  id: number;
  title: string;
  description: string;
  destination: string;
  country: string;
  imageUrl?: string;
  price: number;
  /** Duration in days */
  duration: number;
  maxCapacity: number;
  availableSpots: number;
  startDate: string;
  endDate: string;
  featured: boolean;
  rating: number;
  reviewCount: number;
  includes: string[];
  createdAt: string;
}

export interface CreateTripRequest {
  title: string;
  description: string;
  destination: string;
  country: string;
  imageUrl?: string;
  price: number;
  duration: number;
  maxCapacity: number;
  startDate: string;
  endDate: string;
  featured: boolean;
  includes: string[];
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  tripId: number;
  userId?: number;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  numberOfPeople: number;
  totalPrice: number;
  status: BookingStatus;
  specialRequests?: string;
  trip?: Trip;
  createdAt: string;
}

export interface CreateBookingRequest {
  tripId: number;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  /** @minimum 1 */
  numberOfPeople: number;
  specialRequests?: string;
}

export type UpdateBookingStatusRequestStatus =
  (typeof UpdateBookingStatusRequestStatus)[keyof typeof UpdateBookingStatusRequestStatus];

export const UpdateBookingStatusRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusRequest {
  status: UpdateBookingStatusRequestStatus;
}

export type AdminStatsPopularDestinationsItem = {
  destination: string;
  count: number;
};

export interface AdminStats {
  totalTrips: number;
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  totalRevenue: number;
  popularDestinations: AdminStatsPopularDestinationsItem[];
}

export type GetTripsParams = {
  destination?: string;
  minPrice?: number;
  maxPrice?: number;
  featured?: boolean;
};